Strange zeros being appended to... However fixed main issues with outputstream.

+++ Java/LFS/src/lamp/util/ByteUtil.java
@@ -69,7 +69,7 @@ public final class ByteUtil
 		{
 			if(primitiveList.get(primListIndex) != null)
 			{
-				array[primListIndex] = (byte)primitiveList.get(primListIndex);
+				array[primListIndex] = primitiveList.get(primListIndex);
 			}
 		}
